% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateStates.R
\name{aggregateStates}
\alias{aggregateStates}
\title{merge segments that are all in a row, but avoid slopping across chroms}
\usage{
aggregateStates(HMM, only3 = FALSE, verbose = FALSE, BPPARAM = SerialParam())
}
\arguments{
\item{HMM}{a ChromHMM GRanges}

\item{only3}{keep only `name`, `thick`, `itemRgb` as mcols? (FALSE)}

\item{verbose}{prattle endlessly about which contig we're on? (FALSE)}

\item{BPPARAM}{BiocParallel object (default is SerialParam())}
}
\value{
HMM but without unnecessary segmentations (or mcols)
}
\description{
merge segments that are all in a row, but avoid slopping across chroms
}
