% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorHMM.R
\name{colorHMM}
\alias{colorHMM}
\title{add (new) colors to a ChromHMM track}
\usage{
colorHMM(HMM, cols = NULL, how = c("MNEMONIC", "STATE", "NUMBER"))
}
\arguments{
\item{HMM}{a GRanges from an ChromHMM segmentation}

\item{cols}{a data.frame of colors and associated states}

\item{how}{how to match the HMM states to those in cols? ("MNEMONIC")}
}
\value{
same HMM but with better colors
}
\description{
add (new) colors to a ChromHMM track
}
