% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donutPlot.R
\name{donutPlot}
\alias{donutPlot}
\title{simple function for a donut plot}
\usage{
donutPlot(vec, colorScheme, minpct = 7, leg = FALSE, legTitle = "")
}
\arguments{
\item{vec}{vector of items}

\item{colorScheme}{named vector of colors, one per item type}

\item{minpct}{lower bound to label a slice of the donut}

\item{leg}{legend? (FALSE; no legend)}

\item{legTitle}{title for legend? ("")}
}
\value{
a plot
}
\description{
simple function for a donut plot
}
