% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{GenomicSegmentation}
\alias{GenomicSegmentation}
\title{A subclass of GRanges with a trackLine slot, similar to UCSCData}
\usage{
GenomicSegmentation(gr, ...)
}
\arguments{
\item{gr}{a GRanges or UCSCData object, as from rtracklayer::import.bed}

\item{...}{additional arguments, currently ignored}
}
\value{
a GenomicSegmentation object, very similar to UCSCData
}
\description{
And also a friendly constructor for such things
}
