% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{GenomicSegmentationList}
\alias{GenomicSegmentationList}
\title{Like GRangesList, but with a trackLine slot, similar to UCSCData}
\usage{
GenomicSegmentationList(grl, ...)
}
\arguments{
\item{grl}{a GRangesList, with names that become trackLine names}

\item{...}{additional arguments, currently ignored}
}
\value{
a GenomicSegmentationList object
}
\description{
And also a friendly constructor for such things
}
