% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConditions.R
\name{getConditions}
\alias{getConditions}
\title{list all the conditions found in a BED file of a stacked HMM}
\usage{
getConditions(stacked)
}
\arguments{
\item{stacked}{the stacked HMM}
}
\value{
the conditions found in the stacked HMM (via state:condition)
}
\description{
list all the conditions found in a BED file of a stacked HMM
}
