% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getState.R
\name{getState}
\alias{getState}
\title{pull out ranges from a GenomicSegmentation or similar}
\usage{
getState(gs, state = "Bivalent")
}
\arguments{
\item{gs}{the GenomicSegmentation or GRanges or similar}

\item{state}{the state of interest (default is "Bivalent")}
}
\value{
that subset of the GenomicSegmentation (or similar)
}
\description{
trivial function but still handy
}
