% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStates.R
\name{getStates}
\alias{getStates}
\title{list all the states found in a BED file of an HMM, stacked or otherwise}
\usage{
getStates(HMM, asTable = FALSE)
}
\arguments{
\item{HMM}{the HMM}

\item{asTable}{return a table of states? (FALSE)}
}
\value{
depending on the value of asTable, either a table or a list
}
\description{
list all the states found in a BED file of an HMM, stacked or otherwise
}
