% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importAndSimplifyHMM.R
\name{importAndSimplifyHMM}
\alias{importAndSimplifyHMM}
\title{import and simplify a REMC or ENCODE ChromHMM track}
\usage{
importAndSimplifyHMM(
  x,
  genome = "mm10",
  species = NULL,
  cols = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{x}{a BED filename with a ChromHMM segmentation in it}

\item{genome}{a genome associated with the segmentation (default is mm10)}

\item{species}{a species for the segmentation (default: guess from genome)}

\item{cols}{a color key for simplifying ChromHMMs (pass to simplifyHMM)}

\item{verbose}{be verbose? (FALSE)}

\item{...}{arguments (`only3`, `BPPARAM`) for aggregateStates}
}
\value{
a GRanges of aggregated, simplified HMM states
}
\description{
import and simplify a REMC or ENCODE ChromHMM track
}
