% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAndExportStackedHMM.R
\name{makeAndExportStackedHMM}
\alias{makeAndExportStackedHMM}
\title{convenience function for dumping sets of chromHMM "contrasts" as stacked HMMs}
\usage{
makeAndExportStackedHMM(HMMs, trackName, BPPARAM = SerialParam())
}
\arguments{
\item{HMMs}{a GRangesList of HMMs}

\item{trackName}{what to name the track (will also generate BED filename)}

\item{BPPARAM}{a SerialParam(), by default}
}
\value{
status of the export operation
}
\description{
convenience function for dumping sets of chromHMM "contrasts" as stacked HMMs
}
