% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFrac.R
\name{makeFrac}
\alias{makeFrac}
\title{convenience function for generating donut plots}
\usage{
makeFrac(vec, minpct = 7)
}
\arguments{
\item{vec}{vector of items}

\item{minpct}{lower bound to label a slice}
}
\value{
a data.frame
}
\description{
convenience function for generating donut plots
}
