% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regionFrac.R
\name{regionFrac}
\alias{regionFrac}
\title{convenience function for generating donut plots of count proportions}
\usage{
regionFrac(gs, minpct = 1, what = NULL)
}
\arguments{
\item{gs}{a GRanges or GenomicSegmentation with states}

\item{minpct}{lower bound to label a slice}

\item{what}{optional indicator}
}
\value{
a data.frame
}
\description{
convenience function for generating donut plots of count proportions
}
