% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackedPlot.R
\name{stackedPlot}
\alias{stackedPlot}
\title{simple function for a circular barchart of state transitions in a stackedHMM}
\usage{
stackedPlot(stacked, text = NULL, colorScheme = NULL, legend = FALSE)
}
\arguments{
\item{stacked}{a StackedHMM or similar (a GRanges of one is fine)}

\item{text}{text to put in the middle of the plot (NULL)}

\item{colorScheme}{named vector of colors (default is simplified REMC)}

\item{legend}{legend? (FALSE; no legend)}
}
\value{
a plot
}
\description{
simple function for a circular barchart of state transitions in a stackedHMM
}
