% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackHMMs.R
\name{stackHMMs}
\alias{stackHMMs}
\title{build a poor man's stacked HMM}
\usage{
stackHMMs(
  HMMs,
  djHMM = NULL,
  cols = NULL,
  simple = TRUE,
  only3 = FALSE,
  verbose = FALSE,
  BPPARAM = SerialParam()
)
}
\arguments{
\item{HMMs}{a GRangesList of HMMs}

\item{djHMM}{optional disjoint HMM (else one will be assembled)}

\item{cols}{optional color key (else a default simpleCols will be used)}

\item{simple}{simplify? (default is TRUE, you're on your own if it's FALSE)}

\item{only3}{keep only `name`, `thick`, `itemRgb` as mcols? (FALSE)}

\item{verbose}{be verbose? (FALSE)}

\item{BPPARAM}{BiocParallel object (default is SerialParam())}
}
\value{
a GRanges of stacked HMM states (one or more per segment)
}
\description{
build a poor man's stacked HMM
}
