% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widthFrac.R
\name{widthFrac}
\alias{widthFrac}
\title{convenience function for generating donut plots}
\usage{
widthFrac(gs, minpct = 7)
}
\arguments{
\item{gs}{a GRanges or GenomicSegmentation with states}

\item{minpct}{lower bound (width-based) to label a state}
}
\value{
a data.frame
}
\description{
convenience function for generating donut plots
}
