% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/donutPlot.R
\name{donutPlot}
\alias{donutPlot}
\title{simple function for a donut plot of segment width fractions}
\usage{
donutPlot(gs, colorScheme = NULL, ..., leg = FALSE, legTitle = "")
}
\arguments{
\item{gs}{a GenomicSegmentation or similar}

\item{colorScheme}{named vector of colors (default is simple REMC 25 state)}

\item{...}{additional arguments to pass to widthFrac}

\item{leg}{legend? (FALSE; no legend)}

\item{legTitle}{title for legend? ("")}
}
\value{
a plot
}
\description{
by default, use widthFrac to determine how much of each state is represented
}
