% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplifyHMM.R
\name{simplifyHMM}
\alias{simplifyHMM}
\title{convenience function for adding simplified names & colors to a ChromHMM track}
\usage{
simplifyHMM(HMM, cols = NULL, how = c("MNEMONIC", "STATE", "NUMBER"))
}
\arguments{
\item{HMM}{a GRanges containing a ChromHMM track}

\item{cols}{color table (will be loaded if not provided)}

\item{how}{one of `MNEMONIC` (default), `STATE`, or `NUMBER`}
}
\value{
the same ChromHMM track, but simplified and colored simply
}
\description{
convenience function for adding simplified names & colors to a ChromHMM track
}
\seealso{
colorHMM
}
